VERY IMPORTANT FIX
My fellow stegglang programmers, the future of stegglang is in jeopardy. A recent review of the file `README.md` states that "Steggy is strong dinoswoar and needs only two number sto keep track  of and modify all of his data there are two numbers you can change in memory steggpoint and steggpoint2
```addsteggpoint```/```addsteggpoint2```
to add steggpoint one
```addsteggpoint```/```addsteggpoint2```
to lose steggpoint one
```whatissteggpoint```/```whatissteggpoint2```". This information can be viewed on the official stegglang foundation github repository (at profsucrose/stegglang). However, there is a crucial mistake in the given prompt above. To lose one steggpoint, one must type the keywords `losesteggpoint` or `losesteggpoint2` to lose a steggpoint. This is a major mistake on behalf of the stegglang foundation, and we apologize for any hair-ripping, monitor-punching, or corner-crying that has occurred because of this typo. This fork must be accepted immediately by the president of the stegglang foundation and steggy fanclub, professor sucrose. This mistake must be fixed as soon as possible. As billions of new programmers view this documentation every day, giving false information could corrupt the new age of programmers and hackers from using next-generation stegglang. I ask for the stegglang foundation president and founder, Professor Sucrose, to accept this pull request immediately to fix this error.
   Sincerely,
Steggy LastName